+++ R/install-python.R
@@ -15,26 +15,56 @@
 #' use_virtualenv("my-environment")
 #' ```
 #'
-#' @param version The version of Python to install.
+#' @param version The version of Python to install. If the patch level is
+#'   ommitted, then the latest patch level is automatically selected.
 #'
 #' @param list Boolean; if set, list the set of available Python versions?
 #'
 #' @param force Boolean; force re-installation even if the requested version
 #'   of Python is already installed?
 #'
+#' @param create_virtualenv Boolean; Automatically create a virtualenv with
+#'   the python installation.
+#'
+#' @param ... Passed on to `[virtualenv_create()]` if `create_virtualenv` is
+#'   `TRUE`
+#'
 #' @export
-install_python <- function(version,
+install_python <- function(version = "3.8",
                            list = FALSE,
-                           force = FALSE)
+                           force = FALSE,
+                           create_virtualenv = TRUE,
+                           ...)
 {
   # resolve pyenv path
   pyenv <- pyenv_find()
   if (!file.exists(pyenv))
     stop("could not locate 'pyenv' binary")
 
+  valid_versions <- pyenv_list(pyenv = pyenv)
+
   # if list is set, then list available versions instead
   if (identical(list, TRUE))
-    return(pyenv_list(pyenv = pyenv))
+    return(valid_versions)
+
+  version <- as.character(version)
+
+  if (!(version %in% valid_versions))
+    tryCatch({
+      # accept versions like "3.8", and automatically select the latest patchlevel.
+      # If we error here, just proceed and let pyenv raise the error.
+      # We do this manually here until https://github.com/pyenv/pyenv/issues/2145 is resolved.
+      valid_versions <-
+        valid_versions[startsWith(valid_versions, version)]
+      version. <- paste0(version, ".")
+      patchlevels <- lapply(valid_versions, function(v)
+        suppressWarnings(as.integer(str_drop_prefix(v, version.))))
+      version_w_patchlevel <-
+        paste0(version., max(unlist(patchlevels), na.rm = TRUE))
+      if (version_w_patchlevel %in% valid_versions)
+        version <- version_w_patchlevel
+    },
+    error = function(e) NULL)
 
   # install the requested package
   status <- pyenv_install(version, force, pyenv = pyenv)
@@ -48,7 +78,8 @@ install_python <- function(version,
     pip_install(python, "virtualenv")
 
   # return path to python
-  python
-
+  if (create_virtualenv)
+    virtualenv_create(python = python, ...)
+  else
+    python
 }
-